requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
Pillow==9.4.0
pystray==0.19.4
requests==2.28.1
six==1.16.0
urllib3==1.26.13